pyproject.toml
[project]
name = "proprosals"
version = "0.1.0"
description = ""
authors = [
    {name = "Phil VanEvery",email = "pvanevery@intaptai.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf (>=1.26.4,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
